Fix issues regarding weight computations

+++ src/bftsmart/consensus/Epoch.java
@@ -451,7 +451,7 @@ public class Epoch implements Serializable {
      * @param value The value in question
      * @return total weights from which this process received the specified value
      */
-    public synchronized int countWriteWeigths(byte[] value) {
+    public synchronized double countWriteWeigths(byte[] value) {
         return countWeigths(writeSetted,sumWeightsWrite, write, value);
     }
 
@@ -460,7 +460,7 @@ public class Epoch implements Serializable {
      * @param value The value in question
      * @return total weights from which this process accepted the specified value
      */
-    public synchronized int countAcceptWeigths(byte[] value) {
+    public synchronized double countAcceptWeigths(byte[] value) {
         return countWeigths(acceptSetted,sumWeightsAccept, accept, value);
     }
 
@@ -470,9 +470,9 @@ public class Epoch implements Serializable {
      * @param value Value to count
      * @return Ammount of times that 'value' was find in 'array'
      */
-    private synchronized int countWeigths(boolean[] arraySetted, double[] arrayWeights, byte[][] array, byte[] value) {
+    private synchronized double countWeigths(boolean[] arraySetted, double[] arrayWeights, byte[][] array, byte[] value) {
         if (value != null) {
-            int counter = 0;
+            double counter = 0.0;
             for (int i = 0; i < array.length; i++) {
                 if (arrayWeights != null && arraySetted != null && arraySetted[i] && Arrays.equals(value, array[i])) {
                     counter += arrayWeights[i];
